pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = "present managing bot"
authors = ["Tvister7"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.17.1"
SQLAlchemy = "^1.4.29"
asyncpg = "^0.25.0"
python-dotenv = "^0.19.2"
sqlmodel = "^0.0.5"
aiosqlite = "^0.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
